+++ src/muddybot/__main__.py
@@ -9,7 +9,7 @@ from .files import CONFIG_FILE
 from .twitch_bot import TwitchBot
 from .logger import logger
 
-__version__ = "1.0.4+snapshot25w25d"
+__version__ = "1.0.5+snapshot25w25e"
 
 
 def main() -> int:
@@ -25,7 +25,6 @@ def main() -> int:
                         raise TypeError("The 'token' field must be a string.")
                 except KeyError:
                     logger.error("OAUTH token not set.")
-                    print("OAUTH token not set.")
                     return 1  # Exit with code 1
                 try:
                     channel = config['channel']
@@ -34,24 +33,20 @@ def main() -> int:
                         raise TypeError("The 'channel' field must be a list.")
                     logger.debug(channel)
                 except KeyError:
-                    logger.error("Channel must be set.")
-                    print("You must set the channel in the config file.")
+                    logger.error("Channel must be set in the config file.")
                     return 1  # Exit with code 1
                 try:
                     key = config['key']
                     if not isinstance(key, str):
                         raise TypeError("The 'key' field must be a string.")
                 except KeyError:
-                    logger.critical("Encryption key not found :(")
-                    print("Set the 'key' field in your config file to your encryption key")
+                    logger.critical("Encryption key not found in the config file.")
                     return 1  # Exit with code 1
             except TypeError as error_message:
-                print(error_message)
                 logger.error(error_message)
                 return 1  # Return exit code of 1
     except FileNotFoundError:  # If the config file was not found we fail
         logger.critical("Config file: '%s' not found.", CONFIG_FILE)
-        print(f"Config file: '{CONFIG_FILE}' not found.")
         return 1  # Return with exit code 1
 
     bot = TwitchBot(token, channel, key)  # If we did'nt fail at all we run the twitch bot